make TestDialect static inner class in LimitOffsetFetchClauseIntegrationTests
