+++ src/integrationTest/java/com/mongodb/hibernate/query/select/LimitOffsetFetchClauseIntegrationTests.java
@@ -20,14 +20,26 @@ import static com.mongodb.hibernate.internal.MongoAssertions.assertTrue;
 import static com.mongodb.hibernate.internal.MongoAssertions.fail;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import com.mongodb.hibernate.TestDialect;
+import com.mongodb.hibernate.dialect.MongoDialect;
 import com.mongodb.hibernate.internal.FeatureNotSupportedException;
 import com.mongodb.hibernate.internal.MongoConstants;
 import java.util.Arrays;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
 import org.hibernate.cfg.AvailableSettings;
+import org.hibernate.dialect.Dialect;
+import org.hibernate.engine.jdbc.dialect.spi.DialectResolutionInfo;
+import org.hibernate.engine.spi.SessionFactoryImplementor;
 import org.hibernate.query.SelectionQuery;
 import org.hibernate.query.sqm.FetchClauseType;
+import org.hibernate.sql.ast.SqlAstTranslator;
+import org.hibernate.sql.ast.SqlAstTranslatorFactory;
+import org.hibernate.sql.ast.tree.MutationStatement;
+import org.hibernate.sql.ast.tree.select.SelectStatement;
+import org.hibernate.sql.exec.spi.JdbcOperationQueryMutation;
+import org.hibernate.sql.exec.spi.JdbcOperationQuerySelect;
+import org.hibernate.sql.model.ast.TableMutation;
+import org.hibernate.sql.model.jdbc.JdbcMutationOperation;
 import org.hibernate.testing.orm.junit.DomainModel;
 import org.hibernate.testing.orm.junit.ServiceRegistry;
 import org.hibernate.testing.orm.junit.Setting;
@@ -431,17 +443,20 @@ class LimitOffsetFetchClauseIntegrationTests extends AbstractSelectionQueryInteg
     @ServiceRegistry(
             settings = {
                 @Setting(name = AvailableSettings.QUERY_PLAN_CACHE_ENABLED, value = "true"),
-                @Setting(name = AvailableSettings.DIALECT, value = "com.mongodb.hibernate.TestDialect")
+                @Setting(
+                        name = AvailableSettings.DIALECT,
+                        value =
+                                "com.mongodb.hibernate.query.select.LimitOffsetFetchClauseIntegrationTests$MqlTranslateCacheTestingDialect"),
             })
     class QueryPlanCacheTests extends AbstractSelectionQueryIntegrationTests {
 
         private static final String HQL = "from Book order by id";
 
-        private TestDialect testDialect;
+        private MqlTranslateCacheTestingDialect mqlTranslateCacheTestingDialect;
 
         @BeforeEach
         void beforeEach() {
-            testDialect = (TestDialect) getSessionFactoryScope()
+            mqlTranslateCacheTestingDialect = (MqlTranslateCacheTestingDialect) getSessionFactoryScope()
                     .getSessionFactory()
                     .getJdbcServices()
                     .getDialect();
@@ -453,11 +468,11 @@ class LimitOffsetFetchClauseIntegrationTests extends AbstractSelectionQueryInteg
             getSessionFactoryScope().inTransaction(session -> {
                 var query = session.createSelectionQuery(HQL, Book.class);
                 setQueryOptionsAndQuery(query, isFirstResultSet ? 5 : null, isMaxResultsSet ? 10 : null);
-                var initialSelectTranslatingCount = testDialect.getSelectTranslatingCounter();
+                var initialSelectTranslatingCount = mqlTranslateCacheTestingDialect.getSelectTranslatingCounter();
 
                 query = session.createSelectionQuery(HQL, Book.class);
                 setQueryOptionsAndQuery(query, isFirstResultSet ? 3 : null, isMaxResultsSet ? 6 : null);
-                assertThat(testDialect.getSelectTranslatingCounter()).isEqualTo(initialSelectTranslatingCount);
+                assertThat(mqlTranslateCacheTestingDialect.getSelectTranslatingCounter()).isEqualTo(initialSelectTranslatingCount);
             });
         }
 
@@ -475,16 +490,16 @@ class LimitOffsetFetchClauseIntegrationTests extends AbstractSelectionQueryInteg
         void testCacheInvalidatedDueToQueryOptionsAdded() {
             getSessionFactoryScope().inTransaction(session -> {
                 session.createSelectionQuery(HQL, Book.class).getResultList();
-                var initialSelectTranslatingCount = testDialect.getSelectTranslatingCounter();
+                var initialSelectTranslatingCount = mqlTranslateCacheTestingDialect.getSelectTranslatingCounter();
 
                 session.createSelectionQuery(HQL, Book.class).setFirstResult(1).getResultList();
-                assertThat(testDialect.getSelectTranslatingCounter()).isEqualTo(initialSelectTranslatingCount + 1);
+                assertThat(mqlTranslateCacheTestingDialect.getSelectTranslatingCounter()).isEqualTo(initialSelectTranslatingCount + 1);
 
                 session.createSelectionQuery(HQL, Book.class)
                         .setFirstResult(1)
                         .setMaxResults(5)
                         .getResultList();
-                assertThat(testDialect.getSelectTranslatingCounter()).isEqualTo(initialSelectTranslatingCount + 2);
+                assertThat(mqlTranslateCacheTestingDialect.getSelectTranslatingCounter()).isEqualTo(initialSelectTranslatingCount + 2);
             });
         }
 
@@ -492,10 +507,10 @@ class LimitOffsetFetchClauseIntegrationTests extends AbstractSelectionQueryInteg
         void testCacheInvalidatedDueToQueryOptionsRemoved() {
             getSessionFactoryScope().inTransaction(session -> {
                 session.createSelectionQuery(HQL, Book.class).setFirstResult(10).getResultList();
-                var initialSelectTranslatingCount = testDialect.getSelectTranslatingCounter();
+                var initialSelectTranslatingCount = mqlTranslateCacheTestingDialect.getSelectTranslatingCounter();
 
                 session.createSelectionQuery(HQL, Book.class).getResultList();
-                assertThat(testDialect.getSelectTranslatingCounter()).isEqualTo(initialSelectTranslatingCount + 1);
+                assertThat(mqlTranslateCacheTestingDialect.getSelectTranslatingCounter()).isEqualTo(initialSelectTranslatingCount + 1);
             });
         }
 
@@ -503,11 +518,49 @@ class LimitOffsetFetchClauseIntegrationTests extends AbstractSelectionQueryInteg
         void testCacheInvalidatedDueToDifferentQueryOptions() {
             getSessionFactoryScope().inTransaction(session -> {
                 session.createSelectionQuery(HQL, Book.class).setFirstResult(10).getResultList();
-                var initialSelectTranslatingCount = testDialect.getSelectTranslatingCounter();
+                var initialSelectTranslatingCount = mqlTranslateCacheTestingDialect.getSelectTranslatingCounter();
 
                 session.createSelectionQuery(HQL, Book.class).setMaxResults(20).getResultList();
-                assertThat(testDialect.getSelectTranslatingCounter()).isEqualTo(initialSelectTranslatingCount + 1);
+                assertThat(mqlTranslateCacheTestingDialect.getSelectTranslatingCounter()).isEqualTo(initialSelectTranslatingCount + 1);
             });
+        }
+    }
+
+    public static final class MqlTranslateCacheTestingDialect extends Dialect {
+        private final AtomicInteger selectTranslatingCounter = new AtomicInteger();
+        private final Dialect delegate;
+
+        public MqlTranslateCacheTestingDialect(DialectResolutionInfo info) {
+            super(info);
+            delegate = new MongoDialect(info);
+        }
+
+        @Override
+        public SqlAstTranslatorFactory getSqlAstTranslatorFactory() {
+            return new SqlAstTranslatorFactory() {
+                @Override
+                public SqlAstTranslator<JdbcOperationQuerySelect> buildSelectTranslator(
+                        SessionFactoryImplementor sessionFactory, SelectStatement statement) {
+                    selectTranslatingCounter.incrementAndGet();
+                    return delegate.getSqlAstTranslatorFactory().buildSelectTranslator(sessionFactory, statement);
+                }
+
+                @Override
+                public SqlAstTranslator<? extends JdbcOperationQueryMutation> buildMutationTranslator(
+                        SessionFactoryImplementor sessionFactory, MutationStatement statement) {
+                    return delegate.getSqlAstTranslatorFactory().buildMutationTranslator(sessionFactory, statement);
+                }
+
+                @Override
+                public <O extends JdbcMutationOperation> SqlAstTranslator<O> buildModelMutationTranslator(
+                        TableMutation<O> mutation, SessionFactoryImplementor sessionFactory) {
+                    return delegate.getSqlAstTranslatorFactory().buildModelMutationTranslator(mutation, sessionFactory);
+                }
+            };
+        }
+
+        public int getSelectTranslatingCounter() {
+            return selectTranslatingCounter.get();
         }
     }
 }
